requirements.txt
Django==2.2.2
gitpython